backend/requirements.txt
# Melascope DX Backend Requirements
# Python 3.9+ recommended

# FastAPI and server
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# PyTorch and ML
torch==2.3.0
torchvision==0.18.0
transformers==4.57.6
safetensors==0.4.3
accelerate==0.30.0

# Image processing
Pillow==10.3.0
numpy==1.26.4

# Utilities
pydantic==2.7.0
python-dotenv==1.0.1

# Optional: For better performance
# onnxruntime==1.18.0  # Uncomment if converting to ONNX

httpx==0.27.0
